Commit of the files I have been working on
I created a text file of all the nodes and duplicated your main and made some edits. I made a new function called readData that reads in from the text file and converts it to the correct format to be sent to the createLink function. I had to do some funky stuff to read things into Cstrings but I believe that it all works. The only issue is that when I try to put populate all of the nodes into the graph it is seg faulting for some reason. Haven't had much time to troubleshoot it but a quick run through the debugger was pointing to the createLink function, line 101/102.