Changes to the breadth-first search
Changes I made in main were for not directly calling BFS function to just call a the member function print. This makes a call to BFS to then find the shortest path and store that into the arrays pathTo, distTo, pred where we are able to loop back through the arrays to find are path back to the source and the cost of the distance to the source to destination.